@@ -12,38 +12,35 @@
 		"format": "prettier --write ."
 	},
 	"devDependencies": {
-		"@iconify/json": "^2.2.262",
-		"@iconify/tailwind": "^1.1.3",
-		"@sveltejs/adapter-node": "^5.2.8",
-		"@sveltejs/kit": "^2.7.2",
-		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@date-fns/tz": "^1.2.0",
+		"@iconify/json": "^2.2.291",
+		"@iconify/tailwind": "^1.2.0",
+		"@sveltejs/adapter-node": "^5.2.11",
+		"@sveltejs/kit": "^2.15.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"@types/eslint": "^9.6.1",
-		"@types/node": "^22.7.7",
-		"@typescript-eslint/eslint-plugin": "^8.11.0",
-		"@typescript-eslint/parser": "^8.11.0",
+		"@types/node": "^22.10.4",
+		"@typescript-eslint/eslint-plugin": "^8.19.0",
+		"@typescript-eslint/parser": "^8.19.0",
 		"autoprefixer": "^10.4.20",
-		"eslint": "^9.13.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.46.0",
-		"flowbite": "^2.5.2",
-		"flowbite-svelte": "^0.47.1",
-		"postcss": "^8.4.47",
-		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.7",
-		"prettier-plugin-tailwindcss": "^0.6.8",
-		"svelte": "^5.0.5",
-		"svelte-check": "^4.0.5",
-		"tailwindcss": "^3.4.14",
-		"tslib": "^2.8.0",
-		"typescript": "^5.6.3",
-		"unplugin-icons": "^0.19.3",
-		"vite": "^5.4.9"
-	},
-	"type": "module",
-	"dependencies": {
-		"@date-fns/tz": "^1.1.2",
 		"date-fns": "^4.1.0",
-		"socket.io-client": "^4.8.0",
-		"svelte-markdown": "^0.4.1"
-	}
+		"eslint": "^9.17.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-svelte": "^2.46.1",
+		"flowbite": "^2.5.2",
+		"flowbite-svelte": "^0.47.4",
+		"marked": "^15.0.5",
+		"postcss": "^8.4.49",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.2",
+		"prettier-plugin-tailwindcss": "^0.6.9",
+		"socket.io-client": "^4.8.1",
+		"svelte": "^5.16.1",
+		"svelte-check": "^4.1.1",
+		"tailwindcss": "^3.4.17",
+		"tslib": "^2.8.1",
+		"typescript": "^5.7.2",
+		"vite": "^6.0.7"
+	},
+	"type": "module"
 }
