@@ -1,9 +1,15 @@
 <script lang="ts">
-	export let data: { error?: any };
+	import { page } from '$app/state';
+
+	interface Props {
+		data: { error?: any; noauth?: any; apiUrl?: any };
+		children?: import('svelte').Snippet;
+	}
+	let { data, children }: Props = $props();
 </script>
 
 {#if !data.error}
-	<slot></slot>
+	{@render children?.()}
 {:else}
 	<div class="child:p-2 md:child:p-16 grid grid-cols-1 grid-rows-1 text-center text-white">
 		<h2 class="text-center text-3xl font-bold text-white">Sikertelen API lekérdezés</h2>
