@@ -1,17 +1,23 @@
 <script lang="ts">
 	import { goto } from '$app/navigation';
 	import { page } from '$app/stores';
-	export let data: PageData;
 	import Error from '$lib/error.svelte';
 	import { onMount } from 'svelte';
 	import type { PageData } from '../../routes/ucp/potlekok/$types';
 	import { formatRelative } from 'date-fns';
 	import { locale } from '$lib/time';
-	let multipage = false;
-	export let tipus = '';
-	export let display = '';
-	let handled_potleks: any = [];
-	let pagee = data.page as number;
+	import { get_status_string, get_type_number, get_type_string } from './types';
+	import { Tooltip } from 'flowbite-svelte';
+	let multipage = $state(false);
+	interface Props {
+		data: PageData;
+		tipus: number;
+		display?: string;
+	}
+
+	let { data, tipus, display = '' }: Props = $props();
+	let handled_potleks: any = $state([]);
+	let pagee = $state(data.page as number);
 	function switchPage(mode: 'next' | 'prev') {
 		let url = new URL($page.url);
 		if (mode === 'next') {
@@ -29,9 +35,9 @@
 	}
 	function render() {
 		handled_potleks = [];
-		if (data.potlekok.length > 20 && data.potlekok.length > 0) {
+		if (data.potlekok.length > 10 && data.potlekok.length > 0) {
 			multipage = true;
-			for (let i = (pagee as number) * 20; i < (pagee as number) * 20 + 20; i++) {
+			for (let i = (pagee as number) * 10; i < (pagee as number) * 10 + 10; i++) {
 				if (data.potlekok[i]) {
 					handled_potleks.push(data.potlekok[i]);
 				}
@@ -46,27 +52,39 @@
 </script>
 
 <Error {data}>
-	<div class="grid grid-cols-1 text-center text-black dark:text-white">
-		<a
-			href={`${$page.url.pathname}/upload`}
-			class="from-taxi hover:bg-pos-100 bg-size-200 bg-pos-0 mb-5 ml-5 mr-5 mt-5 block rounded-full bg-gradient-to-r via-teal-400 to-green-600 px-6 py-3 text-center text-xl font-bold text-white drop-shadow-lg transition-all duration-500"
-			>Feltöltés</a
-		>
-		<h1 class="text-5xl font-bold drop-shadow-xl">{display}:</h1>
-		<h2 class="mb-3 text-black dark:text-gray-400">(összesen {data.potlekok.length} darab)</h2>
+	<div
+		class="grid grid-cols-1 content-center items-center justify-center text-center text-black dark:text-white"
+	>
+		<h1 class="mt-5 text-5xl font-bold drop-shadow-xl">{display}:</h1>
+		<div class="m-auto mb-5 flex gap-5">
+			<h2 class="m-auto text-black dark:text-white">
+				Összesen {data.potlekok.length} darab.
+			</h2>
+			<a
+				href={`${$page.url.pathname}/upload`}
+				aria-label="Feltöltés"
+				class="from-taxi hover:bg-pos-100 bg-size-200 bg-pos-0 h-8 w-16 rounded-full bg-gradient-to-r via-teal-400 to-green-600 text-center text-xl font-bold text-white shadow-2xl drop-shadow-lg transition-all duration-500"
+				><span class="icon-[material-symbols--upload] h-full w-full"></span></a
+			>
+			<Tooltip class="bg-slate-500">
+				Új {get_type_string(tipus)} feltöltése
+			</Tooltip>
+		</div>
 		<div class="mb-3 flex flex-auto flex-wrap items-center justify-center gap-3 align-middle">
 			{#if handled_potleks}
 				{#each handled_potleks as potle}
 					<div
 						class="rounded-lg bg-gradient-to-bl p-2 drop-shadow-xl"
-						class:from-teal-500={potle.status === 'elfogadva'}
-						class:to-green-600={potle.status === 'elfogadva'}
-						class:from-amber-500={potle.status === 'elutasítva'}
-						class:to-red-600={potle.status === 'elutasítva'}
-						class:from-cyan-800={potle.status === 'feltöltve'}
-						class:to-gray-700={potle.status === 'feltöltve'}
+						class:from-teal-500={get_status_string(potle.status) === 'elfogadva'}
+						class:to-green-600={get_status_string(potle.status) === 'elfogadva'}
+						class:from-amber-500={get_status_string(potle.status) === 'elutasítva'}
+						class:to-red-600={get_status_string(potle.status) === 'elutasítva'}
+						class:from-cyan-800={get_status_string(potle.status) === 'feltöltve'}
+						class:to-gray-700={get_status_string(potle.status) === 'feltöltve'}
 					>
-						<h1 class="-mb-2 text-2xl font-bold drop-shadow-xl">{potle.status.toUpperCase()}</h1>
+						<h1 class="-mb-2 text-2xl font-bold drop-shadow-xl">
+							{get_status_string(potle.status).toUpperCase()}
+						</h1>
 						<h1 class="text-gray-200 drop-shadow-xl">
 							{formatRelative(new Date(new Date(potle.date).valueOf() - data.offset!), new Date(), {
 								locale
@@ -76,17 +94,17 @@
 							<h1 class="drop-shadow-xl">Megjegyzés: {potle.reason}</h1>
 						{/if}
 
-						{#if tipus === 'leintés'}
+						{#if tipus === get_type_number('leintés')}
 							<div class="flex flex-col xl:flex-row">
 								<a
-									href={`${data.api}/limg?id=${potle.id}&ver=0`}
+									href={`${data.image}/get?id=${potle.img_1}`}
 									target="_blank"
-									on:mouseenter={() => (potle.focus1 = true)}
-									on:mouseleave={() => (potle.focus1 = false)}
+									onmouseenter={() => (potle.focus1 = true)}
+									onmouseleave={() => (potle.focus1 = false)}
 								>
 									<img
 										loading="lazy"
-										src={`${data.api}/limg?id=${potle.id}&ver=0`}
+										src={`${data.image}/get?id=${potle.img_1}`}
 										alt=""
 										class="max-h-xl m-auto max-w-xl py-2 drop-shadow-xl"
 										class:blur={potle.focus1}
@@ -98,14 +116,14 @@
 									{/if}
 								</a>
 								<a
-									href={`${data.api}/limg?id=${potle.id}&ver=1`}
+									href={`${data.image}/get?id=${potle.img_2}`}
 									target="_blank"
-									on:mouseenter={() => (potle.focus2 = true)}
-									on:mouseleave={() => (potle.focus2 = false)}
+									onmouseenter={() => (potle.focus2 = true)}
+									onmouseleave={() => (potle.focus2 = false)}
 								>
 									<img
 										loading="lazy"
-										src={`${data.api}/limg?id=${potle.id}&ver=1`}
+										src={`${data.image}/get?id=${potle.img_2}`}
 										alt=""
 										class="max-h-xl m-auto max-w-xl py-2 drop-shadow-xl"
 										class:blur={potle.focus2}
@@ -119,14 +137,14 @@
 							</div>
 						{:else}
 							<a
-								href={`${data.api}/img?id=${potle.id}`}
+								href={`${data.image}/get?id=${potle.img_1}`}
 								target="_blank"
-								on:mouseenter={() => (potle.focus = true)}
-								on:mouseleave={() => (potle.focus = false)}
+								onmouseenter={() => (potle.focus = true)}
+								onmouseleave={() => (potle.focus = false)}
 							>
 								<img
 									loading="lazy"
-									src={`${data.api}/img?id=${potle.id}`}
+									src={`${data.image}/get?id=${potle.img_1}`}
 									alt=""
 									class="max-h-xl m-auto max-w-xl py-2 drop-shadow-xl"
 									class:blur={potle.focus}
@@ -139,8 +157,8 @@
 							</a>
 						{/if}
 
-						{#if potle.admin}
-							<h1 class="drop-shadow-xl">Kezelte: {potle.admin}</h1>
+						{#if potle.handled_by}
+							<h1 class="drop-shadow-xl">Kezelte: {potle.handled_by}</h1>
 						{/if}
 					</div>
 				{/each}
@@ -154,16 +172,16 @@
 			{#if pagee > 0}
 				<button
 					aria-label="Előző oldal"
-					on:click={() => switchPage('prev')}
+					onclick={() => switchPage('prev')}
 					class="hover:bg-pos-100 bg-size-200 bg-pos-0 rounded-full bg-gradient-to-r from-emerald-500 via-teal-600 to-red-500 text-white duration-300"
 					style="width: calc(5vw*2.5); height: 5vh;"
 					><span class="icon-[solar--map-arrow-left-bold] h-full w-full"></span></button
 				>
 			{/if}
-			{#if Math.ceil(data.potlekok.length / 20) - 1 > pagee}
+			{#if Math.ceil(data.potlekok.length / 10) - 1 > pagee}
 				<button
 					aria-label="Következő oldal"
-					on:click={() => switchPage('next')}
+					onclick={() => switchPage('next')}
 					class="hover:bg-pos-100 bg-size-200 bg-pos-0 rounded-full bg-gradient-to-r from-emerald-500 via-teal-600 to-red-500 text-white duration-300"
 					style="width: calc(5vw*2.5); height: 5vh;"
 					><span class="icon-[solar--map-arrow-right-bold] h-full w-full"></span></button
