@@ -1,12 +1,12 @@
 <script lang="ts">
-	export let data;
+	let { data } = $props();
 </script>
 
 <div class="flex flex-col items-center justify-center">
 	{#each data.cucc as puszedli}
 		{#if data.type === 'szamla'}
 			<h2 class="mt-5 text-center text-xl text-white">
-				{data.cucc.indexOf(puszedli) + 1}. ({puszedli.extra}$):
+				{data.cucc.indexOf(puszedli) + 1}. ({puszedli.price}$):
 			</h2>
 		{:else}
 			<h2 class="mt-5 text-center text-xl text-white">{data.cucc.indexOf(puszedli) + 1}:</h2>
@@ -15,10 +15,13 @@
 			<h2 class="text-center text-gray-400">Megjegyzés: {puszedli.reason}</h2>
 		{/if}
 		{#if data.type === 'leintes'}
-			<img src={`${data.api}/limg?id=${puszedli.id}&ver=0`} alt="kép" />
-			<img src={`${data.api}/limg?id=${puszedli.id}&ver=1`} alt="kép" />
+			<img src={`${data.image}/get?id=${puszedli.img_1}`} alt="kép" />
+			<h2 class="text-gray-400">Kép ID: {puszedli.img_1}</h2>
+			<img src={`${data.image}/get?id=${puszedli.img_2}`} alt="kép" />
+			<h2 class="text-gray-400">Kép ID: {puszedli.img_2}</h2>
 		{:else}
-			<img src={`${data.api}/img?id=${puszedli.id}`} alt="kép" />
+			<img src={`${data.image}/get?id=${puszedli.img_1}`} alt="kép" />
+			<h2 class="text-gray-400">Kép ID: {puszedli.img_1}</h2>
 		{/if}
 		<h2 class="text-gray-400">ID: {puszedli.id}</h2>
 	{/each}
