@@ -2,13 +2,15 @@ use std::env;
 
 use axum::extract::Query;
 use axum::{debug_handler, response::Redirect};
+use base64::engine::general_purpose;
+use base64::Engine;
 use tower_cookies::cookie::time::Duration;
 use tower_cookies::{Cookie, Cookies};
 use url_builder::URLBuilder;
 
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
-pub struct DiscordStuff {
+pub struct DiscordAuth {
     pub api_endpoint: String,
     pub discord_base: String,
     pub discord_id: String,
@@ -19,7 +21,7 @@ pub struct DiscordStuff {
     pub secret_key: String,
 }
 
-pub fn get_discord_envs() -> DiscordStuff {
+pub fn get_discord_envs() -> DiscordAuth {
     let id = env::var("DISCORD_ID")
         .expect("DISCORD_ID .env fájlból betöltése sikertelen. Létre van hozva?");
     let secret = env::var("DISCORD_SECRET")
@@ -32,7 +34,7 @@ pub fn get_discord_envs() -> DiscordStuff {
         .expect("FULL_DOMAIN .env fájlból betöltése sikertelen. Létre van hozva?");
     let secret_key = env::var("SECRET_KEY")
         .expect("SECRET_KEY .env fájlból betöltése sikertelen. Létre van hozva?");
-    DiscordStuff {
+    DiscordAuth {
         api_endpoint: String::from("https://discord.com/api/v10"),
         discord_id: id,
         discord_secret: secret,
@@ -44,21 +46,10 @@ pub fn get_discord_envs() -> DiscordStuff {
     }
 }
 
-pub fn get_auth_url() -> String {
-    let mut ub = URLBuilder::new();
-    let ds = get_discord_envs();
-    ub.set_protocol("https")
-        .set_host(&ds.discord_base.as_str())
-        .add_param("response_type", "code")
-        .add_param("client_id", &ds.discord_id)
-        .add_param("scope", "identify")
-        .add_param("redirect_uri", &ds.redirect_url);
-    ub.build()
-}
-
 #[derive(Deserialize)]
 pub struct Code {
     code: String,
+    state: String,
 }
 
 #[derive(Debug, Deserialize)]
@@ -96,11 +87,40 @@ pub async fn base_callback(Query(query): Query<Code>, cookies: Cookies) -> Redir
     cookies.add(
         Cookie::build(("auth_token", object.access_token))
             .max_age(Duration::seconds(object.expires_in))
+            .http_only(true)
+            .secure(true)
             .domain(ds.domain.clone())
             .path("/")
             .build(),
     );
+    let path = String::from_utf8(general_purpose::STANDARD.decode(query.state).unwrap()).unwrap();
+    let path_full: AuthHomeCode = serde_json::from_str(&path).expect("Nem megy");
+    Redirect::to(&format!("{}{}", &ds.fdomain, path_full.path))
+}
 
-    let red = Redirect::to(&ds.fdomain);
-    red
+fn base_path() -> String {
+    "/ucp".to_string()
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct AuthHomeCode {
+    #[serde(default = "base_path")]
+    path: String,
+}
+
+#[debug_handler]
+pub async fn auth_home(Query(q): Query<AuthHomeCode>) -> Redirect {
+    let mut ub = URLBuilder::new();
+    let state = AuthHomeCode { path: q.path };
+    let state_str = serde_json::to_string(&state).expect("Sikertelen átalakítás");
+    let ds = get_discord_envs();
+    ub.set_protocol("https")
+        .set_host(&ds.discord_base.as_str())
+        .add_param("response_type", "code")
+        .add_param("state", &general_purpose::STANDARD.encode(state_str))
+        .add_param("client_id", &ds.discord_id)
+        .add_param("scope", "identify")
+        .add_param("redirect_uri", &ds.redirect_url);
+    let built_url = ub.build();
+    Redirect::to(&built_url)
 }
