@@ -1,4 +1,9 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
+//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.3
 
-pub mod data;
+pub mod api_keys;
+pub mod bills;
+pub mod hails;
+pub mod images;
+pub mod logs;
 pub mod shorts;
+pub mod supplements;
