@@ -1,4 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
+//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.3
 
 use sea_orm::entity::prelude::*;
 use serde::Serialize;
@@ -6,7 +6,8 @@ use serde::Serialize;
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize)]
 #[sea_orm(table_name = "shorts")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
+    #[sea_orm(primary_key)]
+    pub id: i32,
     pub short: String,
     #[sea_orm(column_type = "Text")]
     pub url: String,
