@@ -1,6 +1,7 @@
-use axum::{response::Redirect, routing::get, Router};
+use std::env;
+
+use axum::{routing::get, Router};
 use dotenvy::dotenv;
-use image::{base_image_get, base_leintes_image_get};
 use socket::InitialData;
 use socketioxide::{
     extract::{Data, SocketRef},
@@ -14,9 +15,9 @@ use tracing_subscriber::FmtSubscriber;
 
 mod auth;
 mod db;
-mod image;
 mod init;
 mod list;
+mod logging;
 mod shorts;
 mod socket;
 mod ucp;
@@ -27,6 +28,21 @@ async fn main() {
     dotenv().expect(".env fájl nem található");
     tracing::subscriber::set_global_default(FmtSubscriber::default()).unwrap();
     let (layer, io) = SocketIo::new_layer();
+    let env_mode = env::var("ENV_MODE");
+    if env_mode.is_err() {
+        panic!("ENV_MODE nincs setelve! production / testing / devel")
+    }
+    if ![
+        "production".to_string(),
+        "testing".to_string(),
+        "devel".to_string(),
+    ]
+    .contains(&env_mode.clone().unwrap())
+    {
+        panic!("ENV_MODE rosszul setelve! production / testing / devel")
+    } else {
+        info!("Running in {} mode", env_mode.unwrap().to_uppercase());
+    }
     init::main();
     io.ns(
         "/",
@@ -35,16 +51,11 @@ async fn main() {
     let app = Router::new()
         .route(
             "/",
-            get(|| async { "SAES API V2 Axum & SQLx használatával" }),
+            get(|| async { "SAES API V2 Axum & Sea-ORM használatával" }),
         )
-        .route(
-            "/auth",
-            get(|| async { Redirect::to(&auth::get_auth_url()) }),
-        )
-        .route("/cb", get(auth::base_callback))
-        .route("/img", get(base_image_get))
+        .route("/auth", get(auth::auth_home))
+        .route("/auth/cb", get(auth::base_callback))
         .route("/list", get(list::base_list_get))
-        .route("/limg", get(base_leintes_image_get))
         .route("/shorts", get(shorts::base_get_shorts))
         .nest("/ucp", ucp::routes())
         .layer(
@@ -55,7 +66,7 @@ async fn main() {
         .layer(TraceLayer::new_for_http())
         .layer(CookieManagerLayer::new());
     // run our app with hyper, listening globally on port 3000
-    info!("Szerver indul");
+    info!("Server runs on :3000");
     let listener = tokio::net::TcpListener::bind("0.0.0.0:3000").await.unwrap();
     axum::serve(listener, app.into_make_service())
         .await
