@@ -1,14 +1,14 @@
 use std::env;
 
 pub struct Apis {
-    pub patrik: String,
-    pub erik: String,
+    pub samt: String,
+    pub sckkapp: String,
 }
 
 pub fn get_api_envs() -> Apis {
-    let patrik = env::var("PATRIK_API")
-        .expect("PATRIK_API .env fájlból betöltése sikertelen. Létre van hozva?");
-    let erik =
-        env::var("ERIK_API").expect("ERIK_API .env fájlból betöltése sikertelen. Létre van hozva?");
-    Apis { patrik, erik }
+    let samt =
+        env::var("SAMT_API").expect("SAMT_API .env fájlból betöltése sikertelen. Létre van hozva?");
+    let sckkapp = env::var("SCKKAPP_API")
+        .expect("SCKKAPP_API .env fájlból betöltése sikertelen. Létre van hozva?");
+    Apis { samt, sckkapp }
 }
