@@ -1,15 +1,17 @@
+use std::env;
+
 use chrono::{Datelike, Duration, Local, NaiveDateTime, NaiveTime};
 
 pub struct Friday {
-    pub last: NaiveDateTime,
-    pub laster: NaiveDateTime,
-    pub next: NaiveDateTime,
+    pub last_friday: NaiveDateTime,
+    pub before_last_friday: NaiveDateTime,
+    pub next_friday: NaiveDateTime,
 }
 
 pub fn get_fridays() -> Friday {
     let input_date = Local::now().date_naive();
     let current_weekday = input_date.weekday().num_days_from_monday();
-    let zaras = NaiveTime::from_hms_opt(20, 0, 0).unwrap();
+    let week_close = NaiveTime::from_hms_opt(20, 0, 0).unwrap();
     // Calculate the number of days until the next Friday (5th day of the week)
     let days_until_friday = (4 + 7 - current_weekday) % 7;
     // If the input date is already Friday, we need to move to the next week's Friday
@@ -23,12 +25,28 @@ pub fn get_fridays() -> Friday {
     let next_friday = input_date + Duration::days(days_until_friday.into());
     let last_friday = next_friday + Duration::days(-7);
     let laster_friday = last_friday + Duration::days(-7);
-    let next_friday_whole = NaiveDateTime::new(next_friday, zaras);
-    let last_friday_whole = NaiveDateTime::new(last_friday, zaras);
-    let laster_friday_whole = NaiveDateTime::new(laster_friday, zaras);
+    let next_friday_whole = NaiveDateTime::new(next_friday, week_close);
+    let last_friday_whole = NaiveDateTime::new(last_friday, week_close);
+    let laster_friday_whole = NaiveDateTime::new(laster_friday, week_close);
     Friday {
-        last: last_friday_whole,
-        laster: laster_friday_whole,
-        next: next_friday_whole,
+        last_friday: last_friday_whole,
+        before_last_friday: laster_friday_whole,
+        next_friday: next_friday_whole,
+    }
+}
+
+#[derive(Debug, PartialEq)]
+pub enum EnvModes {
+    Production,
+    Testing,
+    Devel,
+}
+
+pub fn get_env_mode() -> EnvModes {
+    let mode = env::var("ENV_MODE").unwrap();
+    match mode.as_str() {
+        "production" => EnvModes::Production,
+        "testing" => EnvModes::Testing,
+        &_ => EnvModes::Devel,
     }
 }
