@@ -1,40 +1,32 @@
 use serde::Deserialize;
 
-#[derive(Debug, Deserialize)]
-pub struct BaseImgQuery {
-    pub id: String,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct BaseImgLeintQuery {
-    pub id: String,
-    pub ver: String,
-}
+use super::factions::Factions;
 
 #[derive(Debug, Deserialize)]
 pub struct BaseListQuery {
     pub tipus: String,
     pub driver: String,
+    pub faction: Factions,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct UCPTypeQuery {
-    pub tipus: String,
+    pub tipus: i8,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct UCPTypeExtraQuery {
-    pub tipus: String,
+    pub tipus: i8,
     pub dates: String,
 }
 
 #[derive(Debug, Deserialize)]
-pub struct MVStatQuery {
+pub struct SMStatQuery {
     pub week: String,
 }
 
 #[derive(Debug, Deserialize)]
-pub struct MVItemsQuery {
-    pub tipus: String,
-    pub status: String,
+pub struct SMItemsQuery {
+    pub tipus: i8,
+    pub status: i8,
 }
