@@ -4,7 +4,7 @@ import type { PageServerLoad } from './$types';
 export const load = (async ({ cookies }) => {
 	cookies.delete('auth_token', {
 		path: '/',
-		domain: process.env.NODE_ENV === 'development' ? 'localhost' : 'sckk.hu'
+		domain: process.env.NODE_ENV === 'development' ? 'localhost' : 'samt.hu'
 	});
 	throw redirect(302, '/');
 }) satisfies PageServerLoad;
