@@ -1,31 +1,42 @@
 <script lang="ts">
-	export let data;
-	let hour = new Date().getHours();
-	let greet = '';
-	let reason = '';
+	import { snow } from '$lib/api.js';
 
-	let end = '!';
-	if (hour >= 19) {
-		greet = 'Szép estét';
-	}
-	if (hour >= 13 && hour < 19) {
-		greet = 'Szép délutánt';
-	}
-	if (hour >= 12 && hour < 13) {
-		greet = 'Szép delet';
-	}
-	if (hour >= 8 && hour < 12) {
-		greet = 'Jó reggelt';
-	}
-	if (hour >= 4 && hour < 8) {
-		greet = 'Jó reggelt, bár nem kapsz pótlékot';
-	}
-	if (hour >= 0 && hour < 1) {
-		greet = 'Boldog új évet';
-	}
-	if (hour >= 1 && hour < 4) {
-		greet = 'Miért nem alszol';
-		end = '?';
+	let { data } = $props();
+	let date = new Date();
+	let hour = date.getHours();
+	let greet = $state('');
+	let end = $state('!');
+	if (snow) {
+		greet = 'Kellemes ünnepeket';
+		if (
+			(date.getMonth() === 11 && (date.getDate() === 24 || date.getDate() === 25)) ||
+			(date.getMonth() === 0 && (date.getDate() === 6 || date.getDate() === 7))
+		) {
+			greet = 'Boldog karácsonyt';
+		}
+		if (date.getMonth() === 0 && date.getDate() <= 2) {
+			greet = 'Boldog, és sikerekben gazdag új évet kívánok';
+		}
+	} else {
+		if (hour >= 19 || (hour >= 0 && hour < 2)) {
+			greet = 'Szép estét';
+		}
+		if (hour >= 13 && hour < 19) {
+			greet = 'Szép délutánt';
+		}
+		if (hour >= 12 && hour < 13) {
+			greet = 'Szép delet';
+		}
+		if (hour >= 8 && hour < 12) {
+			greet = 'Jó reggelt';
+		}
+		if (hour >= 4 && hour < 8) {
+			greet = 'Jó reggelt, bár nem kapsz pótlékot';
+		}
+		if (hour >= 2 && hour < 4) {
+			greet = 'Miért nem alszol';
+			end = '?';
+		}
 	}
 	// const sendResponse = async () => {
 	// 	await fetch('/api/upload', {
@@ -60,11 +71,34 @@
 					>
 				</form>
 			{/if} -->
-			{#if !data.layout.am}
+			{#if data.faction == 'SCKK' && data.layout?.taxi}
 				<h2 class="text-xl drop-shadow-lg md:text-2xl">
-					Hívásaid (app+leintés): {data.calls?.app}+{data.calls?.leintes}={Number(data.calls.app) +
-						Number(data.calls?.leintes)}
+					Pozíciód: {data.layout.taxi.positionname}
 				</h2>
+				<h2 class="text-xl drop-shadow-lg md:text-2xl">
+					Műszakod: {data.layout.taxi.shiftname}
+				</h2>
+			{/if}
+			{#if data.faction == 'TOW' && data.layout?.tow}
+				<h2 class="text-xl drop-shadow-lg md:text-2xl">
+					Pozíciód: {data.layout.tow.positionname}
+				</h2>
+				<h2 class="text-xl drop-shadow-lg md:text-2xl">
+					Műszakod: {data.layout.tow.shiftname}
+				</h2>
+			{/if}
+			{#if data.faction === 'SCKK'}
+				{#if data.calls?.app === null}
+					<h2 class="text-xl drop-shadow-lg md:text-2xl">
+						Hívásaid (app nem megy, csak leintés): {data.calls?.leintes}
+					</h2>
+				{:else}
+					<h2 class="text-xl drop-shadow-lg md:text-2xl">
+						Hívásaid (app+leintés): {data.calls?.app}+{data.calls?.leintes}={Number(
+							data.calls?.app
+						) + Number(data.calls?.leintes)}
+					</h2>
+				{/if}
 			{/if}
 			<h2 class="text-xl drop-shadow-lg md:text-2xl">
 				Elfogadott pótlékaid: délelőtti: {data.calls?.potlek.de}, éjszakai: {data.calls?.potlek.du}
