@@ -1,5 +1,5 @@
-import { apiUrl } from '$lib/api';
-import type { RequestHandler } from './$types';
+import { apiUrl } from "$lib/api";
+import type { RequestHandler } from "./$types";
 
 export const POST: RequestHandler = async ({ request, cookies }) => {
 	const body = await request.formData();
@@ -7,27 +7,32 @@ export const POST: RequestHandler = async ({ request, cookies }) => {
 	body.forEach(() => {
 		count++;
 	});
-	const dcauth = cookies.get('auth_token') as string;
-	const tipus = request.headers.get('tip');
-	const dates = request.headers.get('dates');
+	const dcauth = cookies.get("auth_token") as string;
+	const fact = cookies.get("selected_faction") as string;
+	const tipus = request.headers.get("tip");
+	const dates = request.headers.get("dates");
 	const ate = JSON.parse(dates!);
-	if (tipus != 'leintés' || count % 2 === 0) {
+	if (tipus != "leintés" || count % 2 === 0) {
 		if (dcauth) {
-			const mama = await fetch(`${apiUrl}/ucp/items/post?tipus=${tipus}&dates=${ate.toString()}`, {
-				method: 'post',
-				headers: {
-					cookie: dcauth
+			const mama = await fetch(
+				`${apiUrl}/ucp/items/post?tipus=${tipus}&dates=${ate.toString()}`,
+				{
+					method: "post",
+					headers: {
+						cookie: dcauth,
+						faction: fact,
+					},
+					body,
 				},
-				body
-			});
+			);
 			if (mama.status === 406) {
-				return new Response(JSON.stringify({ error: 'toobig' }));
+				return new Response(JSON.stringify({ error: "toobig" }));
 			}
 			const bodi = await mama.json();
 			return new Response(JSON.stringify(bodi));
 		}
 	} else {
-		return new Response(JSON.stringify({ error: 'leintestipik' }));
+		return new Response(JSON.stringify({ error: "leintestipik" }));
 	}
 
 	return new Response(body);
