@@ -1,5 +1,5 @@
-import { addDynamicIconSelectors } from '@iconify/tailwind';
 import flowbite from 'flowbite/plugin';
+import { addDynamicIconSelectors } from '@iconify/tailwind';
 
 /** @type {import('tailwindcss').Config} */
 export default {
@@ -28,7 +28,7 @@ export default {
 			addVariant('child', '& > *');
 			addVariant('child-hover', '& > *:hover');
 		},
-		addDynamicIconSelectors(),
-		flowbite
+		flowbite,
+		addDynamicIconSelectors()
 	]
 };
