@@ -1,12 +1,8 @@
-import { sveltekit } from '@sveltejs/kit/vite';
-import { defineConfig } from 'vite';
-import Icons from 'unplugin-icons/vite';
+import { sveltekit } from "@sveltejs/kit/vite";
+import { defineConfig } from "vite";
 
 export default defineConfig({
 	plugins: [
 		sveltekit(),
-		Icons({
-			compiler: 'svelte'
-		})
-	]
+	],
 });
